Backend/requirements_api.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
langchain-core==0.3.0
langchain-mistralai==0.1.0
langgraph==0.2.0
python-dotenv==1.0.1
PyPDF2==3.0.1
pydantic==2.9.0

